Repo: Outsource ui to its own file including dialogs and prompt
