+++ include/ui.h
@@ -1,0 +1,9 @@
+#ifndef UI_H
+#define UI_H
+
+#include <session.h>
+#include <screen.h>
+
+void fe_safe_file_dialog( Session *s );
+bool fe_quit_dialog( Session *s );
+#endif
